Added get_contactfolder_url helper to generate the Person URL.
